refactor: 컨트롤러를 rest와 view 패키지로 분리
