+++ src/main/java/roomescape/controller/rest/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.api;
+package roomescape.controller.rest;
 
 import java.time.LocalDate;
 import java.util.List;
