+++ src/main/java/roomescape/controller/rest/RoomThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.api;
+package roomescape.controller.rest;
 
 import java.util.List;
 import org.springframework.http.HttpStatus;
