+++ src/test/java/roomescape/controller/rest/ReservationTimeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.api;
+package roomescape.controller.rest;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
